add images on loaclStorage and trying to get but something went wrong
